fixes issues, ui changes
