+++ package.json
@@ -38,10 +38,11 @@
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
     "babel-runtime": "^6.26.0",
-    "bootstrap": "^4.1.3",
+    "bootstrap": "^4.2.1",
     "config": "^2.0.1",
     "core-js": "^2.5.4",
     "express": "^4.16.3",
+    "jquery": "^3.3.1",
     "ngx-cookie-service": "^2.1.0",
     "ngx-logger": "^3.0.5",
     "raven-js": "^3.27.0",
