+++ src/app/components/navbar/navbar.component.html
@@ -1,6 +1,10 @@
-<header>
-  <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <a class="navbar-brand" routerLink="#">JTC</a>
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+  <a class="navbar-brand" routerLink="#">JTC</a>
+  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
+    aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+    <span class="navbar-toggler-icon"></span>
+  </button>
+  <div class="collapse navbar-collapse" id="navbarSupportedContent">
     <ul class="navbar-nav ml-auto">
       <li class="nav-item active" *ngIf="!isLogin">
         <a class="nav-link" routerLink="register">{{'NAVBAR_HEADINGS.USR_CREATE_ACC' | translate}}</a>
@@ -9,11 +13,11 @@
         <a class="nav-link" routerLink="login">{{'NAVBAR_HEADINGS.USR_ACC_LOGIN' | translate}}</a>
       </li>
       <li class="nav-item active" *ngIf="isLogin">
-        <div ngbDropdown class="d-inline-block">
+        <div ngbDropdown>
           <a class="nav-link text-light" ngbDropdownToggle id="dropdownBasic1">{{firstName}}</a>
           <div ngbDropdownMenu aria-labelledby="dropdownBasic1">
             <a routerLink="dashboard" class="dropdown-item">
-                {{'NAVBAR_HEADINGS.USR_ACC_HOME' | translate}}
+              {{'NAVBAR_HEADINGS.USR_ACC_HOME' | translate}}
             </a>
             <a routerLink="dashboard/settings" class="dropdown-item">
               {{'NAVBAR_HEADINGS.USR_ACC_SETTINGS' | translate}}
@@ -25,5 +29,5 @@
         </div>
       </li>
     </ul>
-  </nav>
-  <header>
+  </div>
+</nav>
