+++ src/app/modules/accounts/pages/password-recover/password-recover.component.html
@@ -49,7 +49,7 @@
                 <div class="row mt-3">
                   <div class="col-md-12 form-group text-center">
                     <button type="submit" class="btn btn-lg btn-block btn-primary"
-                            [disabled]="recoverPasswordViewState.isLoading">
+                            [disabled]="recoverPasswordViewState.isLoading || !recoverPasswordFormGroup.valid">
                       {{'FORM_ACTIONS.UPDATE_PASSWORD' | translate}}
                       <span *ngIf="recoverPasswordViewState.isLoading">
                         &nbsp;<i class="fa fa-spinner fa-spin"></i>
