+++ src/app/modules/accounts/pages/password-reset/password-reset.component.html
@@ -29,7 +29,7 @@
                   <div class="row mt-3">
                     <div class="col-md-12 form-group text-center">
                       <button type="submit" class="btn btn-lg btn-block btn-primary"
-                              [disabled]="resetPasswordViewState.isLoading">
+                              [disabled]="resetPasswordViewState.isLoading || !resetPasswordFormGroup.valid">
                         {{'FORM_ACTIONS.USR_ACC_RST_PWD' | translate}}
                         <span *ngIf="resetPasswordViewState.isLoading">
                           &nbsp;<i class="fa fa-spinner fa-spin"></i>
