+++ src/app/modules/dashboard/pages/home/home.component.html
@@ -1,1 +1,9 @@
-<p> Welcome to the dashboard </p>
+<div class="container-fluid container-body">
+  <div class="row">
+    <div class="container col-md-10">
+        <h1 class="text-uppercase">
+            {{'DASH.WELCOME' | translate}}
+          </h1>
+      </div>
+    </div>
+  </div>
