+++ src/assets/i18n/en.json
@@ -12,7 +12,7 @@
     "USR_EMAIL": "Email Address",
     "USR_PWD": "Password",
     "USR_CNF_PWD": "Confirm Password",
-    "USR_PHONE": "Phone Number",
+    "USR_PHONE": "Phone Number (Use Proper Format)",
     "USR_NEW_PASSWORD": "New password"
   },
   "FORM_PLACEHOLDERS": {
@@ -55,6 +55,7 @@
     "USR_ACC_SETTINGS": "Settings"
   },
   "DASH": {
+    "WELCOME": "WELCOME TO DASHBOARD",
     "USR_DRP_DWN_ACTIONS": {
       "SETTINGS": "Settings",
       "ACC_LOGOUT": "Logout"
